backend/requirements.txt
Django>=5.1,<6.0
djangorestframework>=3.15,<4.0
django-cors-headers>=4.6,<5.0
dj-database-url>=2.3,<3.0
psycopg[binary]>=3.2,<4.0
python-dotenv>=1.0,<2.0
supabase>=2.13,<3.0
djangorestframework-simplejwt>=5.4,<6.0
coverage>=7.6,<8.0
openpyxl>=3.1,<4.0
